pyproject.toml
[project]
name = "squakgpt"
version = "0.1.0"
description = "AI-powered sentiment analysis chat app"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy>=2.0.25",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26.0",
    "jinja2>=3.1.3",
    "openai>=1.10.0",
]
